Refactor stop place organisation inserting
Moved mapStopPlaceMaintenanceToInput and changed parameters a bit.
Now insertStopPlaceForScheduledStopPoint handles mapping of maintenance,
so we can use it for inserting maintenance details from E2E tests as well.
